VIITE-3037 Multiple files (BeforeAfter, Discontinuity, LinkStatus, and NodePointType) extracted from liikennevirasto/viite/dao module, and extracted files moved to viite.model package.
